fix precommit
