+++ ctm/supervisors/supervisor_gpt4.py
@@ -1,8 +1,7 @@
 from openai import OpenAI
 
-from utils.exponential_backoff import exponential_backoff
-
-from .supervisor_base import BaseSupervisor
+from ctm.supervisors.supervisor_base import BaseSupervisor
+from ctm.utils.exponential_backoff import exponential_backoff
 
 
 @BaseSupervisor.register_supervisor("gpt4_supervisor")
